requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pycryptodome==3.21.0
pyotp==2.9.0
python-multipart==0.0.9
